progress/pyproject.toml
[project]
name = "Progress"
version = "0.0.1"
description = "A progress tracker for the programming subteam"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.2",
    "jinja2>=3.1.6",
    "pyyaml>=6.0.3",
]
